More MISRA-C changes
